catstream/requirements.txt
Flask>=0.12
matplotlib>=2.1.0
numpy>=1.13
Pillow>=4.3
torch==0.3.0
torchvision==0.2.0
